+++ src/Framework/Container/Container.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Gacela\Framework\Container;
 
+use Closure;
 use Gacela\Framework\Container\Exception\ContainerException;
 use Gacela\Framework\Container\Exception\ContainerKeyNotFoundException;
 use SplObjectStorage;
 
+use function count;
+use function is_callable;
 use function is_object;
 
 final class Container implements ContainerInterface
@@ -20,8 +23,17 @@ final class Container implements ContainerInterface
 
     private SplObjectStorage $factoryServices;
 
-    public function __construct()
+    /** @var array<string,list<Closure>> */
+    private array $servicesToExtend;
+
+    private ?string $currentlyExtending = null;
+
+    /**
+     * @param array<string,list<Closure>> $servicesToExtend
+     */
+    public function __construct(array $servicesToExtend = [])
     {
+        $this->servicesToExtend = $servicesToExtend;
         $this->factoryServices = new SplObjectStorage();
     }
 
@@ -33,6 +45,12 @@ final class Container implements ContainerInterface
     public function set(string $id, $service): void
     {
         $this->services[$id] = $service;
+
+        if ($this->currentlyExtending === $id) {
+            return;
+        }
+
+        $this->extendService($id);
     }
 
     public function has(string $id): bool
@@ -92,5 +110,76 @@ final class Container implements ContainerInterface
             $this->raw[$id],
             $this->services[$id]
         );
+    }
+
+    /**
+     * @psalm-suppress MixedAssignment
+     */
+    public function extend(string $id, Closure $service): Closure
+    {
+        if (!$this->has($id)) {
+            $this->extendLater($id, $service);
+
+            return $service;
+        }
+
+        $factory = $this->services[$id];
+        $extended = $this->generateExtendedService($service, $factory);
+        $this->set($id, $extended);
+
+        return $extended;
+    }
+
+    private function extendLater(string $id, Closure $service): void
+    {
+        if (!isset($this->servicesToExtend[$id])) {
+            $this->servicesToExtend[$id] = [];
+        }
+
+        $this->servicesToExtend[$id][] = $service;
+    }
+
+    /**
+     * @psalm-suppress MissingClosureReturnType,MixedAssignment
+     *
+     * @param mixed $factory
+     */
+    private function generateExtendedService(Closure $service, $factory): Closure
+    {
+        if (is_callable($factory)) {
+            return static function (self $container) use ($service, $factory) {
+                $r1 = $factory($container);
+                $r2 = $service($r1, $container);
+
+                return $r2 ?? $r1;
+            };
+        }
+
+        if (is_object($factory)) {
+            return static function (self $container) use ($service, $factory) {
+                $r = $service($factory, $container);
+
+                return $r ?? $factory;
+            };
+        }
+
+        throw ContainerException::serviceNotExtendable();
+    }
+
+    private function extendService(string $id): void
+    {
+        if (!isset($this->servicesToExtend[$id]) || count($this->servicesToExtend[$id]) === 0) {
+            return;
+        }
+        $this->currentlyExtending = $id;
+
+        foreach ($this->servicesToExtend[$id] as $service) {
+            $extended = $this->extend($id, $service);
+        }
+
+        unset($this->servicesToExtend[$id]);
+        $this->currentlyExtending = null;
+
+        $this->set($id, $extended);
     }
 }